# expand #succ for bignums